reorganizing functions
