+++ syslens-cli/connections.py
@@ -1,4 +1,6 @@
 import paramiko
+from vars import *
+from menu import *
 
 
 def key_based_connect(bastion, username, ssh_key):
@@ -18,3 +20,34 @@ def jump_to_target(client, target, bastion_priv_ip):
         "direct-tcpip", dest_addr, src_addr
     )
     return bastion_channel
+
+
+def main_bastion_connector(host, func):
+    connection = key_based_connect(server, user, ssh_key_filename)
+    bastion_channel = jump_to_target(connection, host, bastion_private_ip)
+    target = paramiko.SSHClient()
+    target.set_missing_host_key_policy(paramiko.AutoAddPolicy())
+    target.connect(
+        host,
+        username=user,
+        key_filename=inner_ssh_key_filename,
+        sock=bastion_channel,
+    )
+    func(target)
+    target.close()
+    bastion_channel.close()
+
+
+def process_hosts(func):
+    host = hosts_selection(server_list)
+    if type(host) is list:
+        for vm in host:
+            print(f"Connecting to: {vm}")
+            main_bastion_connector(vm, func)
+        print("")
+        print("Choose the next option: ")
+    else:
+        print(f"Connecting to: {host}")
+        main_bastion_connector(host, func)
+        print("")
+        print("Choose the next option: ")
